pyproject.toml
[project]
name = "openstadt"
version = "0.1.0"
description = "Open Source Civic Data Platform for German Cities"
authors = [
    {name = "Level09", email = "hello@level09.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    # Enferno core
    "babel>=2.17.0",
    "bcrypt>=4.3.0",
    "bleach>=6.2.0",
    "blinker>=1.9.0",
    "cffi>=1.17.1",
    "click>=8.3.1",
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-caching>=2.3.1",
    "flask-dance>=7.1.0",
    "flask-debugtoolbar>=0.16.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-script>=2.0.6",
    "flask-security-too>=5.7.1",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "flask-babel>=4.0.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "mako>=1.3.10",
    "markupsafe>=3.0.2",
    "orjson>=3.10.18",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pycparser>=2.22",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "qrcode>=8.2",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "speaklater>=1.3",
    "sqlalchemy>=2.0.41",
    "webauthn>=2.6.0",
    "werkzeug>=3.1.4",
    "wtforms>=3.2.1",
    # OpenStadt specific
    "geoalchemy2>=0.15.0",
    "shapely>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.0",
    "geojson>=3.2.0",
]

[project.optional-dependencies]
full = [
    "redis>=7.1.0",
    "celery>=5.5.3",
    "kombu>=5.5.4",
    "amqp>=5.3.1",
    "vine>=5.1.0",
    "meilisearch>=0.33.0",
]
dev = [
    "ruff",
    "pre-commit",
]
wsgi = [
    "uwsgi>=2.0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openstadt"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long, handled by formatter
    "B008",   # do not perform function calls in argument defaults
    "C901",   # complexity
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["openstadt"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
